+++ deltafi-ui/src/components/ActionMetricsTable.vue
@@ -1,12 +1,22 @@
 <template>
   <div>
-    <CollapsiblePanel :header="title" class="metrics-panel table-panel">
-      <DataTable :value="tableData.value" responsive-layout="scroll" striped-rows class="p-datatable-sm p-datatable-gridlines" :loading="loading" sort-field="action_name" :sort-order="1">
-        <template #empty>No {{ title }} available.</template>
-        <template #loading>Loading {{ title }} data. Please wait.</template>
-        <Column header="Action Name" field="action_name" :sortable="true" />
+    <CollapsiblePanel header="Actions" class="metrics-panel table-panel">
+      <DataTable v-model:filters="filters" :value="tableData.value" responsive-layout="scroll" striped-rows class="p-datatable-sm p-datatable-gridlines" :loading="loading" sort-field="action_name" :sort-order="1" filter-display="row">
+        <template #empty>No Action Metrics available.</template>
+        <template #loading>Loading Action Metrics data. Please wait.</template>
+        <Column header="Action Name" field="action_name" :sortable="true">
+          <template #filter="{ filterModel, filterCallback }">
+            <InputText v-model="filterModel.value" type="text" class="p-inputtext-sm p-column-filter" placeholder="Filter by Action Name" @focus="pauseTimer(true)" @blur="pauseTimer(false)" @input="filterCallback()" />
+          </template>
+        </Column>
+        <Column header="Type" field="family_type" :sortable="true" class="type-column">
+          <template #body="{ data }">{{ sentenceCaseString(data.family_type) }}</template>
+          <template #filter="{ filterModel, filterCallback }">
+            <MultiSelect v-model="filterModel.value" type="text" class="p-inputtext-sm deltafi-input-field p-column-filter" placeholder="Filter by Type" :options="familyTypes" @before-show="pauseTimer(true)" @before-hide="pauseTimer(false)" @change="filterCallback()" />
+          </template>
+        </Column>
         <Column v-for="col of columns" :key="col.field" :field="col.field" :header="col.header" :sortable="true" class="metric-column">
-          <template #body="row">{{ formatMatricValue(row) }}</template>
+          <template #body="row">{{ formatMetricValue(row) }}</template>
         </Column>
       </DataTable>
     </CollapsiblePanel>
@@ -16,16 +26,16 @@
 <script setup>
 import Column from "primevue/column";
 import DataTable from "primevue/datatable";
+import InputText from "primevue/inputtext";
+import MultiSelect from "primevue/multiselect";
+import { FilterMatchMode } from "primevue/api";
 import CollapsiblePanel from "@/components/CollapsiblePanel";
 import useUtilFunctions from "@/composables/useUtilFunctions";
-import { computed, defineProps } from "vue";
+import { computed, defineProps, ref, defineEmits } from "vue";
+
+const emit = defineEmits(["pauseTimer"]);
 
 const props = defineProps({
-  family: {
-    type: String,
-    required: false,
-    default: null,
-  },
   actions: {
     type: Object,
     required: true,
@@ -33,18 +43,36 @@ const props = defineProps({
   loading: {
     type: Boolean,
     required: true,
-  },
+  }
 });
 
-const { formattedBytes } = useUtilFunctions();
-const title = computed(() => {
-  let title = "Action Metrics";
-  if (props.family) {
-    const family = props.family.charAt(0).toUpperCase() + props.family.slice(1).toLowerCase();
-    title = `${family} ${title}`;
-  }
-  return title;
+const { sentenceCaseString, formattedBytes } = useUtilFunctions();
+
+const filters = ref({
+  action_name: { value: null, matchMode: FilterMatchMode.CONTAINS },
+  family_type: { value: null, matchMode: FilterMatchMode.IN },
 });
+
+const familyTypes = computed(() => {
+  if (props.actions.length === 0) return [];
+  const familyTypesArray = new Set();
+
+  for (const [actionsKey] of Object.entries(props.actions)) {
+    for (const [actionKey, actionValue] of Object.entries(props.actions[actionsKey])) {
+      if (actionKey == "family_type") {
+        familyTypesArray.add(actionValue);
+      }
+    }
+  }
+  return Array.from(familyTypesArray).sort();
+});
+
+
+// We have to pause the polling timer due to with each time a polling occurs the filter looses focus and clears out if you havent submitted it yet.
+const pauseTimer = (value) => {
+  emit("pauseTimer", value);
+}
+
 const rows = computed(() => {
   if (props.actions.length === 0) return [];
   const actions = props.actions;
@@ -59,7 +87,7 @@ const columns = computed(() => {
   let metricNames = new Set();
   rows.value.forEach((row) => {
     Object.keys(row).forEach((key) => {
-      if (key !== "action_name") metricNames.add(key);
+      if (key !== "action_name" && key !== "family_type") metricNames.add(key);
     });
   });
   return Array.from(metricNames)
@@ -67,21 +95,15 @@ const columns = computed(() => {
     .map((metricName) => {
       return {
         field: metricName,
-        header: metricHeader(metricName),
+        header: sentenceCaseString(metricName),
       };
     });
 });
 const tableData = computed(() => {
   return rows.value.length > 0 && columns.value.length > 0 ? rows : [];
 });
-const metricHeader = (metricName) => {
-  // Metrics names should always be snake case
-  const words = metricName.split("_").map((word) => {
-    return word.charAt(0).toUpperCase() + word.slice(1).toLowerCase();
-  });
-  return words.join(" ");
-};
-const formatMatricValue = (row) => {
+
+const formatMetricValue = (row) => {
   const field = row.column.key;
   const value = row.data[field] || 0;
   if (value === undefined || value === null) {
@@ -93,3 +115,7 @@ const formatMatricValue = (row) => {
   }
 };
 </script>
+
+<style lang="scss">
+@import "@/styles/components/action-metrics-table.scss";
+</style>