+++ deltafi-ui/src/composables/useUtilFunctions.ts
@@ -14,6 +14,7 @@ export default function useUtilFunctions(): {
   formattedBytes: (bytes: number) => string;
   duration: (milliseconds: number, precision: number) => string;
   pluralize: (count: number, singular: string, plural?: string) => string;
+  sentenceCaseString: (stringValue: string) => string;
 } {
   const formatContextData = (contextData: string) => {
     let formattedString = contextData;
@@ -75,6 +76,14 @@ export default function useUtilFunctions(): {
     return uiConfig.useUTC ? dayjs(date).utc().format(format) : dayjs(date).format(format);
   }
 
+  const sentenceCaseString = (stringValue: string) => {
+    // Metrics names should always be snake case
+    const words = stringValue.split("_").map((word) => {
+      return word.charAt(0).toUpperCase() + word.slice(1).toLowerCase();
+    });
+    return words.join(" ");
+  };
+
   return {
     formatTimestamp,
     shortTimezone,
@@ -82,6 +91,7 @@ export default function useUtilFunctions(): {
     formatContextData,
     formattedBytes,
     duration,
-    pluralize
+    pluralize,
+    sentenceCaseString
   };
 }